chore!: Fix compiler warnings
This commit will be considered as breaking because of the API and ABI
changes.

ABI breakage: Some of the fields have been changed from int to size_t.

API breakage: The current API was poorly designed and allows negative
values as parameters to some of the methods.

Take `TextEditor::MoveDown` as an example. By supplying the method with
a negative value, you can make the cursor move up instead.

Any code that relies on/exploits this quirk will be broken.

I'm not gonna prescribe any solution right now, as I Want to focus on
making the codebase more maintainable to me.

This commit only address the compiler warnings emitted from clang, there
will be more breaking changes of same nature later.